requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.37.4
    # via matplotlib
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.6.1
    # via -r requirements.in
numpy==1.23.3
    # via
    #   contourpy
    #   matplotlib
packaging==21.3
    # via matplotlib
pillow==9.2.0
    # via matplotlib
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   matplotlib
pytz==2022.4
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
websocket-client==1.4.1
    # via -r requirements.in
